+++ language/es/knockout.php
@@ -1,0 +1,24 @@
+<?php 
+
+return array(
+
+	/*
+	|--------------------------------------------------------------------------
+	| Administrator Language Lines for knockout
+	|--------------------------------------------------------------------------
+	|
+	| es - Spanish
+	|
+	*/
+
+	'delete_active_item' => '¿Está seguro de que desea eliminar este ítem? Esta operación es irreversible.',
+	'saving' => 'Guardando...',
+	'saved' => 'Ítem guardado.',
+	'deleting' => 'Eliminando...',
+	'deleted' => 'Ítem eliminado.',
+	'character_left' => ' carácter restante',
+	'characters_left' => ' caracteres restantes',
+	'no_results' => 'No se han encontrado resultados',
+	'select_options' => 'Seleccione Algunas Opciones',
+
+);