+++ language/tr/knockout.php
@@ -1,0 +1,22 @@
+<?php
+
+return array(
+
+/*
+|--------------------------------------------------------------------------
+| Administrator Language Lines for knockout
+|--------------------------------------------------------------------------
+|
+| tr - Turkish
+|
+*/
+
+'delete_active_item' => 'Silmek istediğinizden emin misiniz? Bu işlem geri döndürülemez.',
+'saving' => 'Kaydediyor...',
+'saved' => 'Kaydedildi.',
+'deleting' => 'Siliniyor...',
+'deleted' => 'Silindi.',
+'character_left' => ' karakter kaldı',
+'characters_left' => ' karakter kaldı',
+
+);